+++ index.js
@@ -15,6 +15,10 @@ const Bluebird = extendBluebird( require( 'bluebird' ) )
 const SEMVER_PATCH = 'patch'
 	, SEMVER_MINOR = 'minor'
 	, SEMVER_MAJOR = 'major'
+	, SEMVER_PRE_MAJOR = 'premajor'
+	, SEMVER_PRE_MINOR = 'preminor'
+	, SEMVER_PRE_PATCH = 'prepatch'
+	, SEMVER_PRE_RELEASE = 'prerelease'
 	;
 
 function extendInquirer( inquirer ) {
@@ -56,10 +60,10 @@ let logger = ( function() {
 	me.info = info;
 	me.error = error;
 	me.warn = warn;
-	me.log = log;
 	me.line = line;
 	me.command = command;
 	me.title = title;
+	me.out = out;
 
 	// Implementation
 
@@ -104,22 +108,6 @@ let logger = ( function() {
 		out( combine( m, prefix( '>' ), chalk.cyan, newline ) );
 	}
 
-	function log( error, x ) {
-		if ( x === null || x === undefined )
-			return;
-		if ( x.indexOf( '\n' ) > -1 ) {
-			_.each( x.split( '\n' ), data => log( error, data ) );
-			return;
-		} else {
-			out( '| ' );
-			if ( x.trim().length === 0 ) {
-				line( true );
-			} else {
-				out( combine( x.trim(), chalk.bold, error ? chalk.red : chalk.white, newline ) );
-			}
-		}
-	}
-
 	function title( m ) {
 		let hr = _.repeat( '-', m.length );
 		info( hr );
@@ -156,20 +144,20 @@ async function getAppVersion() {
 }
 
 async function validateGitRepository() {
-	let handleError = ( message ) => {
+	let handleError = ( message, command = null ) => {
 		return ( err ) => {
-			throw err.code === 128 ? new ProcedureError( message ) : err;
+			throw err.code === 128 ? new ProcedureError( message, command ) : err;
 		};
 	};
-	await execute( 'git status', false ).catch( handleError( 'Not a valid Git repository.' ) );
+	await execute( 'git status', false ).catch( handleError( 'Not a valid Git repository. To create a new repository, use:', 'git init' ) );
 	await execute( 'git rev-parse HEAD', false ).catch( handleError( 'Git repository seems to be empty, please commit some changes first.' ) );
-	await execute( 'git rev-parse master', false ).catch( handleError( 'Missing branch master, please create it on the first commit of the repository.' ) );
-	await execute( 'git rev-parse develop', false ).catch( handleError( 'Missing branch develop, please branch it off master and use that for development.' ) );
+	await execute( 'git rev-parse master', false ).catch( handleError( 'Missing branch master, please create it on the first commit of the repository:', 'git branch master <first_commit_id>' ) );
+	await execute( 'git rev-parse develop', false ).catch( handleError( 'Missing branch develop, please branch it off master and use that for development:', 'git checkout -b develop master' ) );
 }
 
 async function validateNpmPackage() {
 	if ( !fs.existsSync( 'package.json' ) )
-		throw new ProcedureError( `Folder doesn't seem to contain a valid NPM package.` );
+		throw new ProcedureError( `Folder doesn't seem to contain a valid NPM package. To create a new package, use:`, 'npm init' );
 }
 
 async function getLatestTag() {
@@ -205,12 +193,12 @@ async function countUntrackedFiles() {
 	return execute( 'git ls-files --exclude-standard --others', false ).then( countLines );
 }
 
-async function countDiffCommits() {
-	return execute( `git rev-list --count --right-only master...develop`, false ).then( parseInt );
+async function countDiffCommits( from ) {
+	return execute( `git rev-list --count --right-only ${ from }...HEAD`, false ).then( parseInt );
 }
 
-async function countDiffFiles() {
-	return execute( 'git diff --name-only master..develop', false ).then( countLines );
+async function countDiffFiles( from ) {
+	return execute( `git diff --name-only ${ from }...HEAD`, false ).then( countLines );
 }
 
 function countLines( contents ) {
@@ -234,41 +222,154 @@ async function getRemoteRepositories() {
 }
 
 async function npmTest() {
+
+	let log = ( error, x, lineEnding = false ) => {
+		if ( x === null || x === undefined )
+			return;
+		let lines = x.split( '\n' ),
+			style = chalk.bold,
+			newline = false;
+		for ( let line of lines ) {
+			if ( newline ) {
+				logger.line( true );
+				logLine();
+			}
+			logger.out( style( line ) );
+			newline = true;
+		}
+	};
+	let logLine = () => logger.out( chalk.cyan( '    ' ) );
+
 	return new Bluebird( ( resolve, reject ) => {
-		let test = child_process.spawn( /^win/.test( process.platform ) ? 'npm.cmd' : 'npm', [ 'test' ] );
-		test.stdout.on( 'data', ( data ) => logger.log( false, data.toString() ) );
-		test.stderr.on( 'data', ( data ) => logger.log( true, data.toString() ) );
+
+		logger.info( 'Testing NPM package: ', 'npm test -- --color'  );
+		logLine();
+
+		let test = child_process.spawn( /^win/.test( process.platform ) ? 'npm.cmd' : 'npm', [ 'test', '--', '--color' ] );
+		test.stdout.on( 'data', data => log( false, data.toString() ) );
+		test.stderr.on( 'data', data => log( true, data.toString() ) );
 		test.on( 'close', ( code ) => {
-			logger.log( code !== 0, '=> ' + code );
-			logger.line();
+			logger.line( true );
 			code === 0 ? resolve() : reject();
 		} );
+
 	} );
 }
 
-async function askVersionType( currentVersion ) {
-	let questions = [
-		{
-			name: 'type',
-			type: 'list',
-			message: 'Please select versioning type:',
-			choices: [
-				SEMVER_PATCH,
-				SEMVER_MINOR,
-				SEMVER_MAJOR
-			]
-		},
-		{
-			name: 'confirm',
-			type: 'confirm',
-			message: ( answers ) => `This will update this module to version: ${ semver.inc( currentVersion, answers.type ) }. Confirm? `,
-			default: true
-		}
-	];
+async function askVersionType( currentVersion, diffFiles ) {
 
-	return inquirer
-		.prompt( questions )
-		.then( ( answers ) => answers.confirm ? answers.type : null );
+	let isPrerelease = currentVersion.indexOf( '-' ) > -1;
+	// let major = semver.major( currentVersion );
+	let minor = semver.minor( currentVersion );
+	let patch = semver.patch( currentVersion );
+
+	let choice = ( n, v ) => { return { name : n, value: v }; };
+	let semverFormat = ( name, type, identifier = '' ) => `${ name } ( ${ chalk.cyan.bold( semver.inc( currentVersion, type, identifier ) ) } )`;
+
+	if ( isPrerelease ) {
+
+		let prereleaseType = ( patch === 0 ? ( minor === 0 ? SEMVER_MAJOR : SEMVER_MINOR ) : SEMVER_PATCH );
+		let isAlpha = isPrerelease && currentVersion.indexOf( '-alpha.' ) > -1;
+		let isBeta = isPrerelease && currentVersion.indexOf( '-beta.' ) > -1;
+		let isRC = isPrerelease && currentVersion.indexOf( '-rc.' ) > -1;
+
+		let choices = [];
+
+		if ( diffFiles === 0 ) {
+			if ( isAlpha )
+				choices.push( choice( semverFormat( 'Switch to Beta', SEMVER_PRE_RELEASE, 'beta' ), [ SEMVER_PRE_RELEASE, 'beta' ] ) );
+			if ( isBeta || isAlpha )
+				choices.push( choice( semverFormat( 'Switch to Release Candidate', SEMVER_PRE_RELEASE, 'rc' ), [ SEMVER_PRE_RELEASE, 'rc' ] ) );
+			if ( isRC )
+				choices.push( choice( semverFormat( 'Release this release candidate', prereleaseType ), [ prereleaseType ] ) );
+		} else {
+
+			if ( isAlpha )
+				choices.push( choice( semverFormat( 'New alpha version', SEMVER_PRE_RELEASE ), [ SEMVER_PRE_RELEASE ] ) );
+			else if ( isBeta )
+				choices.push( choice( semverFormat( 'New beta version', SEMVER_PRE_RELEASE ), [ SEMVER_PRE_RELEASE ] ) );
+			else if ( isRC )
+				choices.push( choice( semverFormat( 'New release candidate', SEMVER_PRE_RELEASE ), [ SEMVER_PRE_RELEASE ] ) );
+
+			if ( !isAlpha && !isBeta && !isRC )
+				choices.push( choice( semverFormat( 'Switch to Alpha', SEMVER_PRE_RELEASE, 'alpha' ), [ SEMVER_PRE_RELEASE, 'alpha' ] ) );
+
+			if ( !isBeta && !isRC )
+				choices.push( choice( semverFormat( 'Switch to Beta', SEMVER_PRE_RELEASE, 'beta' ), [ SEMVER_PRE_RELEASE, 'beta' ] ) );
+
+			if ( !isRC ) {
+				choices.push( choice( semverFormat( 'Switch to Release Candidate', SEMVER_PRE_RELEASE, 'rc' ), [ SEMVER_PRE_RELEASE, 'rc' ] ) );
+				choices.push( choice( semverFormat( `Complete prerelease ${ chalk.yellow( '(should use RC)' ) }`, prereleaseType ), [ prereleaseType ] ) );
+			}
+
+		}
+
+		return inquirer.prompt( {
+			name: 'value',
+			type: 'list',
+			message: 'You are currently in a prerelease version. How do you want to proceed?',
+			choices: choices
+		} ).then( answers => answers.value );
+
+	} else {
+
+		let askReleaseType = () => {
+			let versionChoice = ( d, v ) => choice( semverFormat( d, v ), v );
+			let prerelease = 'Prerelease version...';
+			let choices = [
+				versionChoice( 'New patch version', SEMVER_PATCH ),
+				versionChoice( 'New minor version', SEMVER_MINOR ),
+				versionChoice( 'New major version', SEMVER_MAJOR ),
+				prerelease
+			];
+			return inquirer.prompt( {
+				name: 'value',
+				type: 'list',
+				message: 'Start a new release:',
+				choices: choices
+			} ).then( answers => answers.value === prerelease ? askPrereleaseType() : [ answers.value ] );
+		};
+
+		let askPrereleaseType = () => {
+			let preversionChoice = ( d, v ) => choice( semverFormat( d, v ), v );
+			let option_back = '< Back';
+			let choices = [
+				option_back,
+				preversionChoice( 'New pre-patch version', SEMVER_PRE_PATCH ),
+				preversionChoice( 'New pre-minor version', SEMVER_PRE_MINOR ),
+				preversionChoice( 'New pre-major version', SEMVER_PRE_MAJOR )
+			];
+			return inquirer.prompt( {
+				name: 'value',
+				type: 'list',
+				message: 'Start a new prerelease:',
+				choices: choices,
+				default: 1
+			} ).then( answers => answers.value === option_back ? askReleaseType() : askPrereleaseIdentifier( answers.value ) );
+		};
+
+		let askPrereleaseIdentifier = ( prereleaseType ) => {
+			let identifierChoice = ( d, i ) => choice( semverFormat( d, prereleaseType, i ), i );
+			let option_back = '< Back';
+			let choices = [
+				option_back,
+				identifierChoice( 'Alpha', 'alpha' ),
+				identifierChoice( 'Beta', 'beta' ),
+				identifierChoice( 'Release Candidate', 'rc' )
+			];
+			return inquirer.prompt( {
+				name: 'value',
+				type: 'list',
+				message: 'Choose prerelease type:',
+				choices: choices,
+				default: 1
+			} ).then( answers => answers.value === option_back ? askPrereleaseType() : [ prereleaseType, answers.value ] );
+		};
+
+		return askReleaseType();
+
+	}
+
 }
 
 async function askForChangelog( versionType, versionNumber ) {
@@ -285,11 +386,22 @@ async function askForChangelog( versionType, versionNumber ) {
 			type: 'editor',
 			message: 'Please write the contents of the update',
 			when: ( answers ) => versionType === SEMVER_MAJOR || answers.change,
-			default: `## v${ versionNumber } (${ moment().format( 'YYYY/MM/DD' ) })\n- Entry 1\n- Entry 2`
+			default: `## v${ versionNumber } (${ moment().format( 'YYYY/MM/DD HH:mm' ) })\n- Entry 1\n- Entry 2`
 		}
 	];
 
 	return inquirer.prompt( questions ).then( answers => answers.entry );
+
+}
+
+async function askForConfirmation( versionNumber ) {
+
+	return inquirer.prompt( {
+		name: 'confirm',
+		type: 'confirm',
+		message: `Are you sure you wish to update your package to version ${ chalk.cyan( versionNumber ) } ? ${ chalk.bold.yellow( 'This action cannot be easily undone.' ) }`,
+		default: false
+	} ).then( answers => answers.confirm );
 
 }
 
@@ -306,7 +418,7 @@ function writeChangelogEntry( entry ) {
 
 async function activate() {
 
-	let LINE_LENGHT = 40;
+	let LINE_LENGHT = 45;
 	let log = ( m, c, level = 'info' ) => {
 		let doLog = logger[ level ];
 		if ( m === null || m === undefined )
@@ -367,16 +479,16 @@ async function activate() {
 		// There are some files yet to be commited => throw exception
 		throw new ProcedureError( 'Repository not clean, please commit all your files before proceeding:', 'git commit -a' );
 
-	let VERSION = await getLatestTag(), tagFound = false;
-	if ( VERSION ) {
-		log( `Latest tag found:`, VERSION );
+	let LAST_TAG = await getLatestTag(), tagFound = false;
+	if ( LAST_TAG ) {
+		log( `Latest tag found:`, LAST_TAG );
 		tagFound = true;
 	} else {
-		VERSION = '0.0.0';
+		LAST_TAG = '0.0.0';
 		log( `Latest tag found:`, '---', 'warn' );
 	}
 
-	if ( tagFound && !semver.valid( VERSION ) )
+	if ( tagFound && !semver.valid( LAST_TAG ) )
 		throw new ProcedureError( 'Invalid tag found according to SEMVER. Please tag your releases using semver.' );
 
 	let PACKAGE_VERSION = await getPackageVersion();
@@ -385,29 +497,45 @@ async function activate() {
 	if ( !semver.valid( PACKAGE_VERSION ) )
 		throw new ProcedureError( 'Invalid package version according to SEMVER. Please package your releases using semver.' );
 
+	let VERSION;
 	if ( !tagFound ) {
 		VERSION = PACKAGE_VERSION;
 	} else {
-		let sv = semver( VERSION ), sp = semver( PACKAGE_VERSION );
+		let sv = semver( LAST_TAG ), sp = semver( PACKAGE_VERSION );
 		if ( sv.major !== sp.major || sv.minor !== sp.minor || sv.patch !== sp.patch )
 			throw new ProcedureError( `Version mismatched, please tag version ${ PACKAGE_VERSION } on your Git repository:`, `git tag ${ PACKAGE_VERSION } <commit_id>` );
-		VERSION = PACKAGE_VERSION;
+		VERSION = LAST_TAG;
 	}
+
+	let match = LAST_TAG.match( /-(alpha|beta|rc)\.([0-9])+$/ );
+	let IS_PRERELEASE_VERSION = match !== null;
+	let [ PRERELEASE_TYPE, PRERELEASE_NUMBER ] = match ? [ match[1], match[2] ] : [ null, null ];
+
+	if ( IS_PRERELEASE_VERSION )
+		log( `Prerelease information:`, `${ PRERELEASE_TYPE } ${ parseInt( PRERELEASE_NUMBER ) }`, 'info' );
 
 	let UNTRACKED = await countUntrackedFiles();
 	log( `Untracked files detected:`, UNTRACKED, UNTRACKED > 0 ? 'warn' : 'info' );
 
-	let DIFF_COMMITS = await countDiffCommits();
-	log( `Commits since last release:`, DIFF_COMMITS, DIFF_COMMITS === 0 ? 'warn' : 'info' );
+	let DIFF_COMMITS = await countDiffCommits( tagFound ? LAST_TAG : 'master' );
+	log( `Commits since ${ VERSION }:`, DIFF_COMMITS, DIFF_COMMITS === 0 && !IS_PRERELEASE_VERSION ? 'warn' : 'info' );
 
-	if ( DIFF_COMMITS === 0 )
+	let DIFF_MASTER_COMMITS = await countDiffCommits( 'master' );
+	if ( DIFF_MASTER_COMMITS !== DIFF_COMMITS )
+		log( `Commits since last stable release:`, DIFF_MASTER_COMMITS, DIFF_MASTER_COMMITS === 0 && !IS_PRERELEASE_VERSION ? 'warn' : 'info' );
+
+	if ( DIFF_COMMITS === 0 && !IS_PRERELEASE_VERSION )
 		// There are no commits between master and develop -> throw exception
 		throw new ProcedureError( 'No commits detected since last version.' );
 
-	let DIFF_FILES = await countDiffFiles();
-	log( `Files changed since last release:`, DIFF_FILES, DIFF_FILES === 0 ? 'warn' : 'info' );
+	let DIFF_FILES = await countDiffFiles( tagFound ? LAST_TAG : 'master' );
+	log( `Files changed since ${ VERSION }:`, DIFF_FILES, DIFF_FILES === 0 && !IS_PRERELEASE_VERSION ? 'warn' : 'info' );
 
-	if ( DIFF_FILES === 0 )
+	let DIFF_MASTER_FILES = await countDiffFiles( 'master' );
+	if ( DIFF_MASTER_FILES !== DIFF_FILES )
+		log( `Files changed since last stable release:`, DIFF_MASTER_COMMITS, DIFF_MASTER_COMMITS === 0 && !IS_PRERELEASE_VERSION ? 'warn' : 'info' );
+
+	if ( DIFF_FILES === 0 && !IS_PRERELEASE_VERSION )
 		// There are no changes between master and develop -> throw exception
 		throw new ProcedureError( 'No changes detected since last version.' );
 
@@ -421,8 +549,9 @@ async function activate() {
 
 	logger.line();
 
-	logger.command( 'npm test' );
-	await npmTest().catch( err => {
+	await npmTest().then( () => {
+		logger.info( 'All tests passed.' );
+	}, err => {
 		throw new ProcedureError( 'Tests failed.' );
 	} );
 
@@ -447,8 +576,14 @@ async function activate() {
 	// TODO: Show the list of commits that would be added. Should be disabled by default.
 	// git log master..develop --oneline
 
+	if ( !fs.existsSync( '.gitignore' ) )
+		logger.warn( '.gitignore file missing, it is suggested to create it before committing unwanted files.' );
 	if ( !fs.existsSync( 'CHANGELOG.md' ) )
-		logger.warn( 'Changelog file missing, it is suggested to create it.' );
+		logger.warn( 'CHANGELOG.md file missing, it is suggested to create it before a public release.' );
+	if ( !fs.existsSync( 'README.md' ) )
+		logger.warn( 'README.md file missing, it is suggested to create it before a public release.' );
+	if ( !fs.existsSync( 'LICENSE' ) )
+		logger.warn( 'LICENSE file missing, it is suggested to create it before a public release.' );
 
 	//
 	// ----------------------------------------------------
@@ -462,25 +597,31 @@ async function activate() {
 
 	logger.line();
 
-	let VERSION_TYPE = await askVersionType( VERSION );
-	if ( !VERSION_TYPE )
-		throw new ProcedureError( 'Operation aborted by the user.' );
+	let [ VERSION_TYPE, PRERELEASE_IDENTIFIER ] = await askVersionType( VERSION, DIFF_FILES );
 
-	let NEXT_VERSION = semver.inc( VERSION, VERSION_TYPE );
+	let NEXT_VERSION = semver.inc( VERSION, VERSION_TYPE, PRERELEASE_IDENTIFIER );
+
+	let IS_UNSTABLE = _.includes( [ SEMVER_PRE_PATCH, SEMVER_PRE_MINOR, SEMVER_PRE_MAJOR, SEMVER_PRE_RELEASE ], VERSION_TYPE );
 
 	let CHANGELOG = await askForChangelog( VERSION_TYPE, NEXT_VERSION );
 
+	if ( CHANGELOG )
+		fs.writeFileSync( 'CHANGELOG.md.draft', CHANGELOG, 'UTF-8' );
+
 	//
-	// From here we start modifying the Git repository.
-	// Should somehow rollback on failure.
+	// ----------------------------------------------------
+	// Confirm section
+	// ----------------------------------------------------
+	// We ask the user for confirmation.
+	// This is the last step before we proceed to modify
+	//   the repository.
 	//
-	// Flow:
-	// - Create new branch 'releases/vX' for this release.
-	// - Finalize CHANGELOG.md
-	// - Update package.json
-	// - Switch to master and merge 'releases/vX'
-	// - Tag this version
-	// - Switch to develop and merge 'releases/vX'
+
+	logger.line();
+
+	let PROCEED = await askForConfirmation( NEXT_VERSION );
+	if ( !PROCEED )
+		throw new ProcedureError( 'Operation aborted by the user.' );
 
 	//
 	// ----------------------------------------------------
@@ -494,6 +635,14 @@ async function activate() {
 	// TODO: We should however try our best to rollback
 	//   the changes if something happens.
 	//
+	// Flow:
+	// - Create new branch 'releases/vX' for this release.
+	// - Finalize CHANGELOG.md
+	// - Update package.json
+	// - Switch to master and merge 'releases/vX'
+	// - Tag this version
+	// - Switch to develop and merge 'releases/vX'
+
 
 	logger.line( true );
 
@@ -505,23 +654,33 @@ async function activate() {
 
 		logger.info( 'Changelog updated.' );
 		logger.line();
-
-		await execute( `git checkout -b ${ RELEASE_BRANCH }` );
-		await execute( `git add CHANGELOG.md` );
-		await execute( `git commit -m "Updated changelog for v${ NEXT_VERSION }"` );
-	} else {
-		await execute( `git checkout -b ${ RELEASE_BRANCH }` );
 	}
 
-	await execute( `npm version ${ VERSION_TYPE } --git-tag-version=false` );
+	if ( !IS_UNSTABLE )
+		await execute( `git checkout -b ${ RELEASE_BRANCH }` );
+
+	if ( CHANGELOG ) {
+		await execute( `git add CHANGELOG.md` );
+		await execute( `git commit -m "Updated changelog for v${ NEXT_VERSION }"` );
+		await execute( `rm CHANGELOG.md.draft` );
+	}
+
+	await execute( `npm version ${ NEXT_VERSION } --git-tag-version=false` );
 	await execute( `git add package.json` );
 	await execute( `git commit -m "${ NEXT_VERSION }"` );
-	await execute( `git checkout master` );
-	await execute( `git merge --no-ff ${ RELEASE_BRANCH }` );
-	await execute( `git tag ${ RELEASE_TAG }` );
-	await execute( `git checkout develop` );
-	await execute( `git merge --no-ff ${ RELEASE_BRANCH }` );
-	await execute( `git branch -d ${ RELEASE_BRANCH }` );
+
+	if ( !IS_UNSTABLE ) {
+		await execute( `git checkout master` );
+		await execute( `git merge --no-ff ${ RELEASE_BRANCH }` );
+		await execute( `git tag ${ RELEASE_TAG }` );
+
+		await execute( `git checkout develop` );
+		await execute( `git merge --no-ff ${ RELEASE_BRANCH }` );
+
+		await execute( `git branch -d ${ RELEASE_BRANCH }` );
+	} else {
+		await execute( `git tag ${ RELEASE_TAG }` );
+	}
 
 	//
 	// ----------------------------------------------------
@@ -549,7 +708,11 @@ async function activate() {
 		logger.info( `To synchronize your changes to the ${ rep } Git repository, use:`, `git push ${ rep } master develop ${ RELEASE_TAG }` );
 	}
 
-	logger.info( `To publish your changes to the npm repository, use:`, 'npm publish' );
+	if ( IS_UNSTABLE ) {
+		logger.info( `To publish your unstable changes to the npm repository, use:`, 'npm publish --tag=unstable' );
+	} else {
+		logger.info( `To publish your changes to the npm repository, use:`, 'npm publish --tag=latest' );
+	}
 
 	return true;